Customer Segmentation Project
The task is to help an automobile company predict the groups of customers in a new segment in a new market since the company plans to enter new markets with their existing products. As the new market is similar to the company's existing market, this project will use the current customer data for model training. We will then predict the new customers' categories based on the classifier fitted to the training data.

Based on the data sets provided, it is a classification problem using the random forest tree model, decision tree model as well as the OneVSRest Classifier with the highest accuracy of 53.4%.

Data source: https://www.kaggle.com/datasets/kaushiksuresh147/customer-segmentation